Convert create scenario to own route instead of modal and split up components to container and views (dumb vs smart
