+++ client/components/Login/LoginModal.jsx
@@ -1,9 +1,5 @@
 import React, { Component, PropTypes } from 'react';
-import { connect } from 'react-redux';
 import { Form, FormGroup, ControlLabel, Col, Button, Modal } from 'react-bootstrap';
-
-// Dispatched actions
-import { loginUser, closeLoginModal } from '../../redux/actionCreators/login-actions';
 
 class LoginModal extends Component {
   constructor(props) {
@@ -76,27 +72,4 @@ LoginModal.propTypes = {
   isLoginOpen: PropTypes.bool,
 };
 
-const mapStateToProps = state => {
-  const { auth, modal } = state;
-  const { isAuthenticated, errorMessage } = auth;
-  const { isLoginOpen, isSignupOpen } = modal;
-
-  return {
-    isLoginOpen,
-    isSignupOpen,
-    isAuthenticated,
-    errorMessage,
-  };
-};
-
-const mapDispatchToProps = dispatch => ({
-  hideLogin() {
-    dispatch(closeLoginModal());
-  },
-
-  sendLogin(creds) {
-    dispatch(loginUser(creds));
-  },
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(LoginModal);
+export default LoginModal;
