+++ client/entry.js
@@ -7,6 +7,7 @@ import store, { history } from './redux/store';
 import AppContainer from './components/AppContainer.jsx';
 import HomeContainer from './components/Home/HomeContainer.jsx';
 import MainContainer from './components/Main/MainContainer.jsx';
+import NewScenarioContainer from './components/NewScenario/NewScenarioContainer.jsx';
 
 const router = (
   <Provider store={store}>
@@ -14,6 +15,7 @@ const router = (
       <Route path="/" component={AppContainer}>
         <IndexRoute component={HomeContainer} />
         <Route path="/main" component={MainContainer} />
+        <Route path="/new-scenario" component={NewScenarioContainer} />
       </Route>
     </Router>
   </Provider>
