+++ client/redux/reducers/modal-reducer.js
@@ -1,6 +1,5 @@
 import { TOGGLE_LOGIN_MODAL } from '../actionCreators/login-actions';
 import { TOGGLE_SIGNUP_MODAL } from '../actionCreators/signup-actions';
-import { TOGGLE_SCENARIO_MODAL } from '../actionCreators/scenario-actions';
 
 const initialState = {
   isLoginOpen: false,
@@ -18,11 +17,6 @@ export const modalReducer = (state = initialState, action) => {
     case TOGGLE_SIGNUP_MODAL:
       return Object.assign({}, state, {
         isSignupOpen: !state.isSignupOpen,
-      });
-
-    case TOGGLE_SCENARIO_MODAL:
-      return Object.assign({}, state, {
-        isScenarioOpen: !state.isScenarioOpen,
       });
 
     default:
